(fix): intentando hacer que funcione el metodo de aceptar receta porque da error en la bbdd

+++ src/main/java/medicamentos/entities/PacienteMedicamento.java
@@ -28,7 +28,7 @@ public class PacienteMedicamento implements Serializable{
 	 	@Id
 	    @GeneratedValue(strategy = GenerationType.IDENTITY) // ID autoincremental
 	    @Column(name = "id_pacientes_medicamentos")
-	    private Long idPacienteMedicamento;
+	    private int idPacienteMedicamento;
 
 
     @ManyToOne
